Fixed printer to format better with constants.
Fixed Node.computeNodeValue() to work with constants (for now it was
just returning "input" which was the input value that was supposed to
be in place of x.

Must fix:
Getting infinity for the fitness of some trees, how?
this.evaluateTree(xValue) is getting infinity errors. So something is
broken with ERCs still :(

Tree's 3rd constructor which works with NodeMaker.
So, fix NodeMaker.makeNode() which won't work with ERCs right now.
Probably have to use regex or sumn?

Get log to print nicer

Need to make: